+++ packages/cli/__mocks__/@averjs/core.js
@@ -1,8 +1,11 @@
 export const mockRun = jest.fn();
+export const mockBuild = jest.fn();
 
 const core = jest.fn().mockImplementation(() => {
   return {
-    run: mockRun
+    config: {},
+    run: mockRun,
+    build: mockBuild
   };
 });