+++ packages/config/lib/configs/renderer.js
@@ -1,4 +1,8 @@
+import path from 'path';
+
 export default () => ({
+  transpileDependencies: [],
+  postcss: {},
   css: {
     extract: false,
     styleResources: {
@@ -6,7 +10,15 @@ export default () => ({
       options: {}
     }
   },
-  purgeCss: false,
+  alias: {
+    '@': path.join(process.env.PROJECT_PATH),
+    '@@': path.resolve(process.env.PROJECT_PATH, '../'),
+    '@components': path.resolve(process.env.PROJECT_PATH, './components'),
+    '@resources': path.resolve(process.env.PROJECT_PATH, './resources'),
+    '@mixins': path.resolve(process.env.PROJECT_PATH, './mixins'),
+    '@pages': path.resolve(process.env.PROJECT_PATH, './pages'),
+    '@vuex': path.resolve(process.env.PROJECT_PATH, './vuex')
+  },
   base: false,
   client: false,
   server: false,