+++ packages/renderer/lib/config/base.js
@@ -1,12 +1,10 @@
 import path from 'path';
-import glob from 'glob-all';
 import webpack from 'webpack';
 import WebpackChain from 'webpack-chain';
 import { VueLoaderPlugin } from 'vue-loader';
-import { warmup } from 'thread-loader';
 import ExtractCssPlugin from 'extract-css-chunks-webpack-plugin';
-import PurgeCssPlugin from 'purgecss-webpack-plugin';
 import StyleLoader from '../utils/style-loader';
+import PerformanceLoader from '../utils/perf-loader';
 import Webpackbar from 'webpackbar';
 import FilesChanged from '../plugins/FilesChanged';
 
@@ -18,14 +16,27 @@ export default class WebpackBaseConfiguration {
     this.cacheDir = path.resolve('node_modules/.cache/averjs');
     
     this.isProd = process.env.NODE_ENV === 'production';
-    this.nodeEnv = process.env.NODE_ENV || 'development';
 
     this.commonRules = [];
 
-    if (!this.isProd) warmup({}, [ 'babel-loader', 'css-loader' ]);
-
     const { webpack } = aver.config;
     this.globalConfig = webpack;
+
+    this.perfLoader = new PerformanceLoader(this.isServer, this.globalConfig);
+    this.perfLoader.warmupLoaders();
+    this.styleLoader = new StyleLoader(this.isServer, this.globalConfig, this.perfLoader);
+  }
+
+  get transpileDeps() {
+    return this.globalConfig.transpileDependencies.map(dep => {
+      if (typeof dep === 'string') {
+        return new RegExp(dep);
+      } else if (dep instanceof RegExp) {
+        return dep;
+      } else {
+        return false;
+      }
+    }).filter(_ => _);
   }
 
   plugins() {
@@ -62,45 +73,23 @@ export default class WebpackBaseConfiguration {
           .end()
         .plugin('module-concatenation')
           .use(webpack.optimize.ModuleConcatenationPlugin);
-            
-      if (this.globalConfig.purgeCss) {
-        this.chainConfig
-          .plugin('purge-css')
-            .use(PurgeCssPlugin, [ {
-              paths: glob.sync([
-                path.resolve(process.env.PROJECT_PATH, './**/*.js'),
-                path.resolve(process.env.PROJECT_PATH, './**/*.vue')
-              ]),
-              whitelistPatterns: [ /^_/ ]
-            } ]);
-      }
     }
   }
 
   alias() {
-    this.chainConfig.resolve
-      .alias
-        .set('@', path.join(process.env.PROJECT_PATH))
-        .set('@@', path.resolve(process.env.PROJECT_PATH, '../'))
-        .set('@components', path.resolve(process.env.PROJECT_PATH, './components'))
-        .set('@resources', path.resolve(process.env.PROJECT_PATH, './resources'))
-        .set('@mixins', path.resolve(process.env.PROJECT_PATH, './mixins'))
-        .set('@pages', path.resolve(process.env.PROJECT_PATH, './pages'))
-        .set('@vuex', path.resolve(process.env.PROJECT_PATH, './vuex'));
+    for (const alias of Object.keys(this.globalConfig.alias)) {
+      this.chainConfig.resolve.alias.set(alias, this.globalConfig.alias[alias]);
+    }
   }
 
   rules() {
-    this.chainConfig.module
+    const vueLoaderRule = this.chainConfig.module
       .rule('vue-loader')
-        .test(/\.vue$/)
-        .use('cache-loader')
-          .loader('cache-loader')
-          .options({
-            cacheDirectory: path.resolve(process.env.PROJECT_PATH, '../node_modules/.cache/cache-loader'),
-            cacheIdentifier: 'vue-loader'
-          })
-          .end()
-        .use('vue-loader')
+        .test(/\.vue$/);
+
+    this.perfLoader.apply(vueLoaderRule, 'vue');
+
+    vueLoaderRule.use('vue-loader')
           .loader('vue-loader')
           .options({
             compilerOptions: {
@@ -134,27 +123,31 @@ export default class WebpackBaseConfiguration {
             cache: true
           });
 
-    this.chainConfig.module
+    const jsRule = this.chainConfig.module
       .rule('js')
         .test(/\.js$/)
-        .include
-          .add(process.env.PROJECT_PATH)
-          .end()
-        .use('cache-loader')
-          .loader('cache-loader')
-          .options({
-            cacheDirectory: path.resolve(process.env.PROJECT_PATH, '../node_modules/.cache/cache-loader'),
-            cacheIdentifier: 'js'
+        .exclude
+          .add(filepath => {
+            // always transpile javascript in vue files
+            if (/\.vue\.js$/.test(filepath)) return false;
+            
+            // transpile project path
+            if (filepath.includes(process.env.PROJECT_PATH)) return false;
+
+            // transpile cache dir
+            if (filepath.includes(this.cacheDir)) return false;
+
+            // check if user wants to transpile it
+            if (this.transpileDeps.some(dep => dep.test(filepath))) return false;
+
+            // Ignore node_modules
+            return /node_modules/.test(filepath);
           })
-          .end()
-        .use('thread-loader')
-          .loader('thread-loader')
-          .options({
-            name: 'js',
-            poolTimeout: !this.isProd ? Infinity : 2000
-          })
-          .end()
-        .use('babel-loader')
+          .end();
+
+    this.perfLoader.apply(jsRule, 'js');
+
+    jsRule.use('babel-loader')
           .loader('babel-loader')
           .options({
             presets: [
@@ -189,16 +182,17 @@ export default class WebpackBaseConfiguration {
         .exclude
           .add(/node_modules/)
           .end()
-        .use('yaml')
-          .loader('json-loader!yaml-loader');
-                
-    const styleLoader = new StyleLoader(this.isServer, this.globalConfig);
+        .use('json-loader')
+          .loader('json-loader')
+          .end()
+        .use('yaml-loader')
+          .loader('yaml-loader');
 
     const cssRule = this.chainConfig.module.rule('css-loader').test(/\.css$/);
-    styleLoader.apply('css', cssRule);
+    this.styleLoader.apply('css', cssRule);
 
     const scssRule = this.chainConfig.module.rule('scss-loader').test(/\.scss$/);
-    styleLoader.apply('scss', scssRule, [ {
+    this.styleLoader.apply('scss', scssRule, [ {
       name: 'sass-loader',
       options: { sourceMap: !this.isProd }
     } ]);
@@ -210,7 +204,8 @@ export default class WebpackBaseConfiguration {
           .loader('url-loader')
           .options({
             limit: 1000,
-            name: '_averjs/fonts/[name].[hash:7].[ext]'
+            name: '_averjs/fonts/[name].[hash:7].[ext]',
+            esModule: false
           });
         
     this.chainConfig.module
@@ -220,7 +215,8 @@ export default class WebpackBaseConfiguration {
           .loader('url-loader')
           .options({
             limit: 1000,
-            name: '_averjs/img/[name].[hash:7].[ext]'
+            name: '_averjs/img/[name].[hash:7].[ext]',
+            esModule: false
           });
         
     this.chainConfig.module
@@ -229,7 +225,8 @@ export default class WebpackBaseConfiguration {
         .use('file-loader')
           .loader('file-loader')
           .options({
-            name: '_averjs/videos/[name].[hash:7].[ext]'
+            name: '_averjs/videos/[name].[hash:7].[ext]',
+            esModule: false
           });
         
     this.chainConfig.module
@@ -238,7 +235,8 @@ export default class WebpackBaseConfiguration {
         .use('file-loader')
           .loader('file-loader')
           .options({
-            name: '_averjs/resources/[name].[hash:7].[ext]'
+            name: '_averjs/resources/[name].[hash:7].[ext]',
+            esModule: false
           });
   }